TypeScript, Code Quality, No Functional Changes
