+++ tgui/packages/tgui/interfaces/Terminal/types.ts
@@ -1,0 +1,25 @@
+/**
+ * @file
+ * @copyright 2023
+ * @author Garash2k (https://github.com/garash2k)
+ * @license ISC
+ */
+
+export type PeripheralData = {
+  card: string,
+  icon: string,
+  label: string,
+  Clown?: boolean,
+  color: any,
+  index: number,
+};
+
+export type TerminalData = {
+  displayHTML: string,
+  TermActive: boolean,
+  windowName: string,
+  fontColor: string,
+  bgColor: string,
+  peripherals: Array<PeripheralData>,
+  doScrollBottom: Boolean
+};
